requirements.txt
# Core Dependencies
aiohttp>=3.8.0
redis>=4.5.0
rq>=1.13.0
rq-scheduler>=0.13.0
psutil>=5.9.0

# API Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
websockets>=12.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pyjwt>=2.8.0

# LLM and AI
openai>=1.0.0
tiktoken>=0.5.0
langchain>=0.1.0
langchain-openai>=0.0.2

# Document Processing
pymupdf>=1.23.0  # Single PDF library with best performance
pypdf2>=3.0.0  # Still needed by some modules
pdfplumber>=0.9.0  # Still needed by some modules

# Search and Knowledge
numpy>=1.24.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2

# Web and Research
beautifulsoup4>=4.12.0
arxiv>=1.4.0
scholarly>=1.7.0
google-search-results>=2.4.0  # For SERPAPI
aiofiles>=23.0.0
requests>=2.31.0
Wikipedia-API>=0.5.4

# Utilities
python-dotenv>=1.0.0
croniter>=1.3.0
schedule>=1.2.0
asyncio>=3.4.3
typing-extensions>=4.5.0

# RunPod (optional, for serverless deployment)
# runpod>=1.0.0httpx==0.27.0
